chore: consolidate colors, theme, and commonStyles into theme.ts
