--- apps/computer-vision/colors.ts
@@ -1,6 +1,0 @@
-const ColorPalette = {
-  primary: '#001A72',
-  strongPrimary: '#020F3C',
-};
-
-export default ColorPalette;
